Refactoring hungryman package to wherefordinner.
Adding profile builds to use cloud native functions and removed rabbitmq
binder.

Enabling redis

Updating Crawler API feign client with new API definition.
