+++ where-for-dinner/where-for-dinner-search-proc/pom.xml
@@ -5,7 +5,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.7.4</version>
+		<version>2.7.5</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.example</groupId>
@@ -15,7 +15,7 @@
 	<description>Where for Dinner search processor</description>
 	<properties>
 		<java.version>11</java.version>
-		<spring-cloud.version>2021.0.3</spring-cloud.version>
+		<spring-cloud.version>2021.0.4</spring-cloud.version>
 		<feign-reactor.version>3.2.1</feign-reactor.version>
 		<snakeyaml.version>1.32</snakeyaml.version>
 	</properties>
@@ -27,15 +27,11 @@
 		<dependency>
   			<groupId>org.springframework.boot</groupId>
   			<artifactId>spring-boot-starter-actuator</artifactId>
-		</dependency>		
+		</dependency>	
 		<dependency>
-            <groupId>org.springframework.cloud</groupId>
-            <artifactId>spring-cloud-stream</artifactId>
-        </dependency>	
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-stream-binder-rabbit</artifactId>
-		</dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-data-redis-reactive</artifactId>
+		</dependency>			
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-starter-function-webflux</artifactId>
@@ -65,13 +61,6 @@
 			<artifactId>reactor-test</artifactId>
 			<scope>test</scope>
 		</dependency>
-		<dependency>
-		    <groupId>org.springframework.cloud</groupId>
-		    <artifactId>spring-cloud-stream</artifactId>
-		    <type>test-jar</type>
-		    <scope>test</scope>
-		    <classifier>test-binder</classifier>
-		</dependency>	
 	</dependencies>
 	<dependencyManagement>
 		<dependencies>
@@ -84,7 +73,40 @@
 			</dependency>
 		</dependencies>
 	</dependencyManagement>
+    <profiles>
+       <profile>
+           <id>streams</id>
+           <activation>
+              <activeByDefault>true</activeByDefault>
+           </activation> 
+           <dependencies>
+			   <dependency>
+		            <groupId>org.springframework.cloud</groupId>
+		            <artifactId>spring-cloud-stream</artifactId>
+		       </dependency>	
+			   <dependency>
+					<groupId>org.springframework.cloud</groupId>
+					<artifactId>spring-cloud-stream-binder-rabbit</artifactId>
+			   </dependency>
+			   <dependency>
+				    <groupId>org.springframework.cloud</groupId>
+				    <artifactId>spring-cloud-stream</artifactId>
+				    <type>test-jar</type>
+				    <scope>test</scope>
+				    <classifier>test-binder</classifier>
+			   </dependency>	
+		   </dependencies>
+       </profile>
+       <profile>
+           <!-- Stubbed for native image build -->
+           <id>native</id>
+           <dependencies>
 
+		   </dependencies>
+           <build>
+           </build>
+       </profile>
+    </profiles>
 	<build>
 		<plugins>
 			<plugin>
