--- where-for-dinner/where-for-dinner-search-proc/src/main/java/com/java/example/tanzu/hungryman/searcher/Searcher.java
@@ -1,11 +1,0 @@
-package com.java.example.tanzu.hungryman.searcher;
-
-import com.java.example.tanzu.hungryman.model.Availability;
-import com.java.example.tanzu.hungryman.model.SearchCriteria;
-
-import reactor.core.publisher.Flux;
-
-public interface Searcher 
-{
-	public Flux<Availability> search(SearchCriteria crit);
-}
