+++ where-for-dinner/where-for-dinner-search-proc/src/main/java/com/java/example/tanzu/wherefordinner/processor/SearchProcessor.java
@@ -1,4 +1,4 @@
-package com.java.example.tanzu.hungryman.processor;
+package com.java.example.tanzu.wherefordinner.processor;
 
 import java.time.Duration;
 
@@ -6,10 +6,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 
-import com.java.example.tanzu.hungryman.model.Availability;
-import com.java.example.tanzu.hungryman.model.SearchCriteria;
-import com.java.example.tanzu.hungryman.processor.cache.HashCache;
-import com.java.example.tanzu.hungryman.searcher.Searcher;
+import com.java.example.tanzu.wherefordinner.model.Availability;
+import com.java.example.tanzu.wherefordinner.model.SearchCriteria;
+import com.java.example.tanzu.wherefordinner.processor.cache.HashCache;
+import com.java.example.tanzu.wherefordinner.searcher.Searcher;
 
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Flux;
@@ -58,26 +58,7 @@ public class SearchProcessor
 								
 								return cache.setHashForKey(key, hash, expiration)
 										.then(Mono.just(avail));
-							});	
-										
-//						return redisTemplate.opsForValue().get(key)
-//							.switchIfEmpty(Mono.just("")) 
-//							.flatMap(value -> 
-//							{
-//								final var hash = generateHash(avail);
-//								
-//								if (StringUtils.hasText(value))
-//									if (hash.equals(value))
-//										return Mono.empty();
-//	
-//								/*
-//								 * Expire the entry after the search window has passed
-//								 */
-//								var expiration = Duration.ofMillis(Math.abs(crit.getEndTime() - System.currentTimeMillis()));
-//								
-//								return redisTemplate.opsForValue().set(key, generateHash(avail), expiration)
-//											.then(Mono.just(avail));
-//							});					
+							});															
 				});
 			});		
 	}
