+++ where-for-dinner/where-for-dinner-search-proc/src/main/java/com/java/example/tanzu/wherefordinner/processor/cache/HashCache.java
@@ -1,4 +1,4 @@
-package com.java.example.tanzu.hungryman.processor.cache;
+package com.java.example.tanzu.wherefordinner.processor.cache;
 
 import java.time.Duration;
