+++ where-for-dinner/where-for-dinner-search-proc/src/main/java/com/java/example/tanzu/wherefordinner/processor/cache/MemoryHashCache.java
@@ -1,14 +1,16 @@
-package com.java.example.tanzu.hungryman.processor.cache;
+package com.java.example.tanzu.wherefordinner.processor.cache;
 
 import java.time.Duration;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
 import reactor.core.publisher.Mono;
 
 @Component
+@Profile("!redis")
 public class MemoryHashCache implements HashCache
 {
 	protected final Map<String, String> cache = new ConcurrentHashMap<>();
