+++ where-for-dinner/where-for-dinner-search-proc/src/main/java/com/java/example/tanzu/wherefordinner/searcher/Searcher.java
@@ -1,0 +1,11 @@
+package com.java.example.tanzu.wherefordinner.searcher;
+
+import com.java.example.tanzu.wherefordinner.model.Availability;
+import com.java.example.tanzu.wherefordinner.model.SearchCriteria;
+
+import reactor.core.publisher.Flux;
+
+public interface Searcher 
+{
+	public Flux<Availability> search(SearchCriteria crit);
+}
