+++ where-for-dinner/where-for-dinner-search-proc/src/test/java/com/java/example/tanzu/wherefordinner/WhereForDinnerSearchProcApplicationTests.java
@@ -1,10 +1,10 @@
-package com.java.example.tanzu.hungryman;
+package com.java.example.tanzu.wherefordinner;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class HungrymanSearchProcApplicationTests {
+class WhereForDinnerSearchProcApplicationTests {
 
 	@Test
 	void contextLoads() {
